api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
PyJWT>=2.9.0
pydantic>=2.10.0
